+++ validator_test.go
@@ -8,30 +8,33 @@ import (
 )
 
 type ValidatorTest struct {
-	auth_email_file *os.File
-	done            chan bool
-	update_seen     bool
+	authEmailFileName string
+	done              chan bool
+	update_seen       bool
 }
 
 func NewValidatorTest(t *testing.T) *ValidatorTest {
 	vt := &ValidatorTest{}
 	var err error
-	vt.auth_email_file, err = ioutil.TempFile("", "test_auth_emails_")
+	f, err := ioutil.TempFile("", "test_auth_emails_")
 	if err != nil {
-		t.Fatal("failed to create temp file: " + err.Error())
+		t.Fatalf("failed to create temp file: %v", err)
 	}
-	t.Logf("%s using temp file %s", t.Name(), vt.auth_email_file.Name())
+	if err := f.Close(); err != nil {
+		t.Fatalf("failed to close temp file: %v", err)
+	}
+	vt.authEmailFileName = f.Name()
 	vt.done = make(chan bool)
 	return vt
 }
 
 func (vt *ValidatorTest) TearDown() {
 	vt.done <- true
-	os.Remove(vt.auth_email_file.Name())
+	os.Remove(vt.authEmailFileName)
 }
 
 func (vt *ValidatorTest) NewValidator(domains []string, updated chan<- bool) func(string) bool {
-	return newValidatorImpl(domains, vt.auth_email_file.Name(), vt.done, func() {
+	return newValidatorImpl(domains, vt.authEmailFileName, vt.done, func() {
 		if vt.update_seen == false {
 			updated <- true
 			vt.update_seen = true
@@ -39,12 +42,18 @@ func (vt *ValidatorTest) NewValidator(domains []string, updated chan<- bool) fun
 	})
 }
 
-// This will close vt.auth_email_file.
 func (vt *ValidatorTest) WriteEmails(t *testing.T, emails []string) {
-	vt.auth_email_file.WriteString(strings.Join(emails, "\n"))
-	if err := vt.auth_email_file.Close(); err != nil {
-		t.Fatal("failed to close temp file " +
-			vt.auth_email_file.Name() + ": " + err.Error())
+	f, err := os.OpenFile(vt.authEmailFileName, os.O_WRONLY, 0600)
+	if err != nil {
+		t.Fatalf("failed to open auth email file: %v", err)
+	}
+
+	if _, err := f.WriteString(strings.Join(emails, "\n")); err != nil {
+		t.Fatalf("failed to write emails to auth email file: %v", err)
+	}
+
+	if err := f.Close(); err != nil {
+		t.Fatalf("failed to close auth email file: %v", err)
 	}
 }
 
@@ -156,5 +165,45 @@ func TestValidatorIgnoreSpacesInAuthEmails(t *testing.T) {
 
 	if !validator("foo.bar@example.com") {
 		t.Error("email should validate")
+	}
+}
+
+func TestValidatorOverwriteEmailListDirectly(t *testing.T) {
+	vt := NewValidatorTest(t)
+	defer vt.TearDown()
+
+	vt.WriteEmails(t, []string{
+		"xyzzy@example.com",
+		"plugh@example.com",
+	})
+	domains := []string(nil)
+	updated := make(chan bool)
+	validator := vt.NewValidator(domains, updated)
+
+	if !validator("xyzzy@example.com") {
+		t.Error("first email in list should validate")
+	}
+	if !validator("plugh@example.com") {
+		t.Error("second email in list should validate")
+	}
+	if validator("xyzzy.plugh@example.com") {
+		t.Error("email not in list that matches no domains " +
+			"should not validate")
+	}
+
+	vt.WriteEmails(t, []string{
+		"xyzzy.plugh@example.com",
+		"plugh@example.com",
+	})
+	<-updated
+
+	if validator("xyzzy@example.com") {
+		t.Error("email removed from list should not validate")
+	}
+	if !validator("plugh@example.com") {
+		t.Error("email retained in list should validate")
+	}
+	if !validator("xyzzy.plugh@example.com") {
+		t.Error("email added to list should validate")
 	}
 }
